+++ chat-widget/src/common/Constants.ts
@@ -65,7 +65,9 @@ export class Constants {
         "application/vnd.microsoft.card.thumbnail",
         "application/vnd.microsoft.card.signin",
         "application/vnd.microsoft.card.oauth",
+        "application/vnd.microsoft.card.video"
     ];
+    public static readonly adaptiveCardContentTypePrefix = "application/vnd.microsoft.card";
     public static readonly maxUploadFileSize = "500000";
     public static readonly imageRegex = /(\.)(jpeg|jpg|jiff|png|gif|bmp|webp)$/i;
     public static readonly audioMediaRegex = /(\.)(aac|aiff|alac|amr|flac|mp2|mp3|pcm|wav|wma)$/i;